% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMBounds2.R
\name{LCMBounds2}
\alias{LCMBounds2}
\title{Finds an exact confindence interval for a linear combination of multinomial probabilities}
\usage{
LCMBounds2(
  Lhat,
  weights,
  samplesizes,
  alpha.upper = 0.025,
  alpha.lower = alpha.upper,
  num.iters = 10,
  power.of.2 = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{Lhat}{Observed value of a linear combination of multinomial probabilities that needs a confidence interval (a scalar).}

\item{weights}{A list of numeric vectors that correspond to the multinomial probabilities.}

\item{samplesizes}{A numeric vector of length K, each entry in the vector cooresponds to the sample sizes for the K multinomial experiements.}

\item{alpha.upper}{The amount of probability outside the upper limit of the confidence interval (this produces a (1-alpha.upper-alpha.lower)\% CI).}

\item{alpha.lower}{The amount of probability outside the lower limit of the confidence interval.}

\item{num.iters}{A scalar which directs the alogrithm on how many random starting locations the optimizer should attempt.  Higher is better, but more costly computationally.}

\item{power.of.2}{A logical tuning parameter.  It is not recommended to change it from the default.}
}
\description{
Finds an exact confindence interval for a linear combination of multinomial probabilities
}
\examples{
Lhat <- 7.8
weightsA <- list(c(0,1,1),c(2,0,3),c(5,3,0))
samplesizes <- c(20,20,20)
LCMBounds2(Lhat,weightsA,samplesizes)

}
