% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possiblesupport.R
\name{possible.support}
\alias{possible.support}
\title{Short Title description}
\usage{
possible.support(weights, samplesizes)
}
\arguments{
\item{weights}{A list of numeric vectors that correspond to the multinomial probabilities.}

\item{samplesizes}{A numeric vector of the multinomial sample sizes.}
}
\description{
Long Description.
}
\examples{
weights <- list(c(0,1,1),c(2,0,3),c(5,3,0))
samplesizes <- c(20,20,20)
possible.support(weights,samplesizes)

}
