% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDirichDraw.R
\name{rDirichDraw}
\alias{rDirichDraw}
\title{Function to draw 'random' probability vectors fixed at a particular L}
\usage{
rDirichDraw(mass, L, bag)
}
\arguments{
\item{mass}{Not sure Yet.}

\item{L}{Not sure Yet.}

\item{bag}{Not sure Yet.}
}
\description{
Function to draw 'random' probability vectors fixed at a particular L
}
