% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_support.R
\name{find_support}
\alias{find_support}
\title{Short Title description}
\usage{
find_support(weights, samplesizes, power.of.2 = TRUE)
}
\arguments{
\item{weights}{A list of numeric vectors that correspond to the multinomial probabilities.}

\item{samplesizes}{A numeric vector of the multinomial sample sizes.}

\item{power.of.2}{logical variable that if set to TRUE will return a support whose length is a power of 2.}
}
\description{
Long Description.
}
\examples{
weights <- list(c(0,1,1),c(2,0,3),c(5,3,0))
samplesizes <- c(20,20,20)
find_support(weights,samplesizes)

}
