% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMBounds.R
\name{LCMBounds}
\alias{LCMBounds}
\title{Finds a Confindence Interval for the Linear Combination of Multinomial Probabilities}
\usage{
LCMBounds(Lhat, weights, samplesizes, alpha.upper = 0.025,
  alpha.lower = alpha.upper, rand.samps = 200, opt.samps = 200,
  mass = 1)
}
\arguments{
\item{Lhat}{Observed value that needs a Confidence Interval (a scalar).}

\item{weights}{A list of numeric vectors that correspond to the multinomial probabilities.}

\item{samplesizes}{A numeric vector of the sample sizes for the K multinomial experiements.}

\item{alpha.upper}{Not sure Yet.}

\item{alpha.lower}{Not sure Yet.}

\item{rand.samps}{Not sure Yet.}

\item{opt.samps}{Not sure Yet.}

\item{mass}{Not sure Yet.}
}
\description{
Finds a Confindence Interval for the Linear Combination of Multinomial Probabilities
}
\examples{
Lhat <- 0.2
weightsA <- list(c(0,1,1),c(2,0,3),c(5,3,0))
samplesizes5 <- c(20,20,20)
LCMBounds(Lhat,weightsA,samplesizes5,rand.samps=20,opt.samps=20)

}
